🔧: Fixing Jest coverage collection and cleaning up config for Jest 22 prep (#5193)
