+++ src/test/mochitest/browser_dbg-reload.js
@@ -20,7 +20,7 @@ async function waitForBreakpoint(dbg, location) {
 add_task(async function() {
   const dbg = await initDebugger("reload/doc-reload.html");
 
-  await waitForSource(dbg, "sjs_code_reload")
+  await waitForSource(dbg, "sjs_code_reload");
   await selectSource(dbg, "sjs_code_reload");
   await waitForSelectedSource(dbg, "sjs_code_reload");
   await addBreakpoint(dbg, "sjs_code_reload", 2);
