--- src/test/tests-framework.js
@@ -1,49 +1,0 @@
-import { startSourceMapWorker, stopSourceMapWorker } from "devtools-source-map";
-
-import {
-  startPrettyPrintWorker,
-  stopPrettyPrintWorker
-} from "../workers/pretty-print";
-
-import {
-  startParserWorker,
-  stopParserWorker,
-  clearSymbols,
-  clearASTs
-} from "../workers/parser";
-import { startSearchWorker, stopSearchWorker } from "../workers/search";
-import { clearDocuments } from "../utils/editor";
-import { getValue } from "devtools-config";
-import { clearHistory } from "./utils/history";
-
-global.jasmine.DEFAULT_TIMEOUT_INTERVAL = 20000;
-global.performance = { now: () => 0 };
-
-function formatException(reason, p) {
-  console && console.log("Unhandled Rejection at:", p, "reason:", reason);
-}
-
-beforeAll(() => {
-  startSourceMapWorker(getValue("workers.sourceMapURL"));
-  startPrettyPrintWorker(getValue("workers.prettyPrintURL"));
-  startParserWorker(getValue("workers.parserURL"));
-  startSearchWorker(getValue("workers.searchURL"));
-  process.on("unhandledRejection", formatException);
-});
-
-afterAll(() => {
-  stopSourceMapWorker();
-  stopPrettyPrintWorker();
-  stopParserWorker();
-  stopSearchWorker();
-  process.removeListener("unhandledRejection", formatException);
-});
-
-afterEach(() => {});
-
-beforeEach(() => {
-  clearASTs();
-  clearSymbols();
-  clearHistory();
-  clearDocuments();
-});
